refactor: reverse protocolmeta dependency so public package wraps internal
Move implementations to internal/protocolmeta and make protocolmeta (public)
re-export from internal, aligning with the established plugin/internal/plugin
convention. Also fix gofmt issues and update E2E golden files for new headers.
