+++ internal/protocolmeta/encode.go
@@ -1,8 +1,10 @@
 package protocolmeta
 
-import public "github.com/scenarigo/scenarigo/protocolmeta"
+import (
+	"net/url"
+)
 
 // EncodeHTTPValue encodes a value for use in HTTP headers using URL path escaping.
 func EncodeHTTPValue(value string) string {
-	return public.EncodeHTTPValue(value)
+	return url.PathEscape(value)
 }
