+++ internal/protocolmeta/normalize.go
@@ -1,8 +1,22 @@
 package protocolmeta
 
-import public "github.com/scenarigo/scenarigo/protocolmeta"
+import (
+	"os"
+	"path/filepath"
+)
 
 // NormalizeScenarioFilepath returns a stable, relative filepath when possible.
 func NormalizeScenarioFilepath(path string) string {
-	return public.NormalizeScenarioFilepath(path)
+	if path == "" || !filepath.IsAbs(path) {
+		return path
+	}
+	wd, err := os.Getwd()
+	if err != nil {
+		return path
+	}
+	rel, err := filepath.Rel(wd, path)
+	if err != nil {
+		return path
+	}
+	return rel
 }
