+++ protocolmeta/encode.go
@@ -1,10 +1,8 @@
 package protocolmeta
 
-import (
-	"net/url"
-)
+import "github.com/scenarigo/scenarigo/internal/protocolmeta"
 
 // EncodeHTTPValue encodes a value for use in HTTP headers using URL path escaping.
 func EncodeHTTPValue(value string) string {
-	return url.PathEscape(value)
+	return protocolmeta.EncodeHTTPValue(value)
 }
