+++ protocolmeta/normalize.go
@@ -1,22 +1,8 @@
 package protocolmeta
 
-import (
-	"os"
-	"path/filepath"
-)
+import "github.com/scenarigo/scenarigo/internal/protocolmeta"
 
 // NormalizeScenarioFilepath returns a stable, relative filepath when possible.
 func NormalizeScenarioFilepath(path string) string {
-	if path == "" || !filepath.IsAbs(path) {
-		return path
-	}
-	wd, err := os.Getwd()
-	if err != nil {
-		return path
-	}
-	rel, err := filepath.Rel(wd, path)
-	if err != nil {
-		return path
-	}
-	return rel
+	return protocolmeta.NormalizeScenarioFilepath(path)
 }
